LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-audio-background.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<audio src="../../content/test.wav" style="width: 300px;" controls></audio>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing that we show a gray background for iOS audio elements.");

const mediaControls = window.internals.shadowRoot(document.querySelector('audio')).lastElementChild;

shouldBecomeEqual("mediaControls.querySelector('.controls-bar') instanceof HTMLDivElement", "true", () => {
    shouldBeEqualToString("getComputedStyle(mediaControls.querySelector('.controls-bar'), ':before').webkitAppearance", "none");
    shouldBeEqualToString("getComputedStyle(mediaControls.querySelector('.controls-bar'), ':before').backgroundColor", "rgb(212, 212, 212)");
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
